requirements.txt
# Web scraping
aiohttp==3.9.5
aiohttp-retry==2.8.3
beautifulsoup4==4.12.3
lxml==5.2.2
chardet==5.2.0

# CLI & progress
rich==13.7.1
tqdm==4.66.4
click==8.1.7

# Data
ujson==5.10.0
python-dotenv==1.0.1

# LLM & API
httpx==0.27.0
openai==1.30.1
pydantic==2.7.1

# Database / vector search
supabase==2.4.6
asyncpg==0.29.0
hnswlib==0.8.0
numpy==2.0.2

# FastAPI
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Utils
tenacity==8.3.0
asyncio-throttle==1.0.2
